! Add about controller to manage contact and about style actions moves various functions from various controllers to a centralized about/contact area
