+++ sources/ElkArte/Controller/Who.php
@@ -276,27 +276,4 @@ class Who extends AbstractController
 		// Any profile fields disabled?
 		$context['disabled_fields'] = isset($modSettings['disabled_profile_fields']) ? array_flip(explode(',', $modSettings['disabled_profile_fields'])) : array();
 	}
-
-	/**
-	 * It prepares credit and copyright information for the credits page or the admin page.
-	 *
-	 * - Accessed by ?action=who;sa=credits
-	 *
-	 * @uses Who language file
-	 * @uses template_credits() sub template in Who.template,
-	 */
-	public function action_credits()
-	{
-		global $context, $txt;
-
-		require_once(SUBSDIR . '/Who.subs.php');
-		Txt::load('Who');
-
-		$context += prepareCreditsData();
-
-		theme()->getTemplates()->load('Who');
-		$context['sub_template'] = 'credits';
-		$context['robot_no_index'] = true;
-		$context['page_title'] = $txt['credits'];
-	}
 }
