+++ sources/ElkArte/Languages/Who/English.php
@@ -128,27 +128,6 @@ $txt['who_topicbyemail'] = 'Email starting a new topic in <a href="%1$s">%2$s</a
 $txt['whotopic_postbyemail'] = 'Email posting in <a href="%1$s">%2$s</a>.';
 $txt['whoall_pm_byemail'] = 'Sending a personal message by email.';
 
-// Credits text
-$txt['credits'] = 'Credits';
-$txt['credits_intro'] = 'ElkArte is 100% free and open-source. We encourage and support an active, open community that accepts contributions from the public.  We want to thank everyone who supported the project by providing code, feedback, bug reports, and opinions, as none of this would have been possible without you.';
-$txt['credits_contributors'] = 'Contributors';
-$txt['credits_and'] = 'and';
-$txt['credits_copyright'] = 'Copyrights';
-$txt['credits_forum'] = 'Forum';
-$txt['credits_addons'] = 'Add-ons';
-$txt['credits_software_graphics'] = 'Software/Graphics';
-$txt['credits_software'] = 'Software';
-$txt['credits_graphics'] = 'Graphics';
-$txt['credits_fonts'] = 'Fonts';
-$txt['credits_groups_contrib'] = 'Contributors';
-$txt['credits_contrib_list'] = 'For a complete list of the many individuals that contributed to the design and implementation of Elkarte, please refer to the official <a href="https://github.com/elkarte/Elkarte/graphs/contributors" target="_blank" class="new_win linkbutton">GitHub contributors.</a>';
-$txt['credits_license'] = 'License';
-$txt['credits_copyright'] = 'Copyright';
-$txt['credits_version'] = 'Version';
-// Replace "English" with the name of this language pack in the string below.
-$txt['credits_groups_translators'] = 'Language Translators';
-$txt['credits_translators_message'] = 'Thank you for your efforts which make it possible for people all around the world to use ElkArte.  For a complete list please refer to the offical <a href="https://www.transifex.com/organization/elkarte/dashboard" target="_blank" class="new_win linkbutton">Transifex contributors.</a>';
-
 // Overrides the already defined strings to get clean results in the table
 $txt['today'] = '%1$s';
 $txt['yesterday'] = '%1$s';
