+++ sources/ElkArte/UserSettingsLoader.php
@@ -502,7 +502,7 @@ class UserSettingsLoader
 		// Check if the account is activated - COPPA first...
 		if ($activation_status === 5)
 		{
-			$context['login_errors'][] = $txt['coppa_no_concent'] . ' <a href="' . getUrl('action', ['action' => 'register', 'sa' => 'coppa', 'member' => $this->settings['id_member']]) . '">' . $txt['coppa_need_more_details'] . '</a>';
+			$context['login_errors'][] = $txt['coppa_no_concent'] . ' <a href="' . getUrl('action', ['action' => 'about', 'sa' => 'coppa', 'member' => $this->settings['id_member']]) . '">' . $txt['coppa_need_more_details'] . '</a>';
 
 			return false;
 		}
