+++ sources/subs/Who.subs.php
@@ -124,59 +124,6 @@ function formatViewers($id, $type)
 	krsort($context['view_members']);
 }
 
-/**
- * This function reads from the database the addons credits,
- * and returns them in an array for display in credits section of the site.
- * The addons copyright, license, title information are those saved from <license>
- * and <credits> tags in package.xml.
- *
- * @return array
- */
-function addonsCredits()
-{
-	$db = database();
-
-	$cache = Cache::instance();
-	$credits = array();
-	if (!$cache->getVar($credits, 'addons_credits', 86400))
-	{
-		$db->fetchQuery('
-			SELECT 
-				version, name, credits
-			FROM {db_prefix}log_packages
-			WHERE install_state = {int:installed_adds}
-				AND credits != {string:empty}
-				AND SUBSTRING(filename, 1, 9) != {string:old_patch_name}
-				AND SUBSTRING(filename, 1, 9) != {string:patch_name}',
-			array(
-				'installed_adds' => 1,
-				'old_patch_name' => 'smf_patch',
-				'patch_name' => 'elk_patch',
-				'empty' => '',
-			)
-		)->fetch_callback(
-			function ($row) use (&$credits) {
-				global $txt;
-
-				$credit_info = Util::unserialize($row['credits']);
-
-				$copyright = empty($credit_info['copyright']) ? '' : $txt['credits_copyright'] . ' &copy; ' . Util::htmlspecialchars($credit_info['copyright']);
-				$license = empty($credit_info['license']) ? '' : $txt['credits_license'] . ': ' . Util::htmlspecialchars($credit_info['license']);
-				$version = $txt['credits_version'] . '' . $row['version'];
-				$title = (empty($credit_info['title']) ? $row['name'] : Util::htmlspecialchars($credit_info['title'])) . ': ' . $version;
-
-				// Build this one out and stash it away
-				$name = empty($credit_info['url']) ? $title : '<a href="' . $credit_info['url'] . '">' . $title . '</a>';
-				$credits[] = $name . (!empty($license) ? ' | ' . $license : '') . (!empty($copyright) ? ' | ' . $copyright : '');
-
-			}
-		);
-
-		$cache->put('addons_credits', $credits, 86400);
-	}
-
-	return $credits;
-}
 
 /**
  * This function determines the actions of the members passed in urls.
@@ -297,7 +244,7 @@ function determineActions($urls, $preferred_prefix = false)
 		else
 		{
 			// Viewing/editing a profile.
-			if ($actions['action'] == 'profile')
+			if ($actions['action'] === 'profile')
 			{
 				// Whose?  Their own?
 				if (empty($actions['u']))
@@ -316,13 +263,13 @@ function determineActions($urls, $preferred_prefix = false)
 				}
 
 				$data[$k] = $txt['who_hidden'];
-				$profile_ids[(int) $actions['u']][$k] = $actions['action'] == 'profile' ? $txt['who_viewprofile'] : $txt['who_profile'];
+				$profile_ids[(int) $actions['u']][$k] = $actions['action'] === 'profile' ? $txt['who_viewprofile'] : $txt['who_profile'];
 			}
 			// Trying to post
-			elseif (($actions['action'] == 'post' || $actions['action'] == 'post2' || $actions['action'] == 'topicbyemail') && empty($actions['topic']) && isset($actions['board']))
+			elseif (($actions['action'] === 'post' || $actions['action'] === 'post2' || $actions['action'] === 'topicbyemail') && empty($actions['topic']) && isset($actions['board']))
 			{
 				$data[$k] = $txt['who_hidden'];
-				if ($actions['action'] == 'topicbyemail')
+				if ($actions['action'] === 'topicbyemail')
 				{
 					$board_ids[(int) $actions['board']][$k] = $txt['who_topicbyemail'];
 				}
@@ -332,7 +279,7 @@ function determineActions($urls, $preferred_prefix = false)
 				}
 			}
 			// A subaction anyone can view... if the language string is there, show it.
-			elseif (isset($actions['sa']) && isset($txt['whoall_' . $actions['action'] . '_' . $actions['sa']]))
+			elseif (isset($actions['sa'], $txt['whoall_' . $actions['action'] . '_' . $actions['sa']]))
 			{
 				$data[$k] = $preferred_prefix && isset($txt[$preferred_prefix . $actions['action'] . '_' . $actions['sa']]) ? $txt[$preferred_prefix . $actions['action'] . '_' . $actions['sa']] : $txt['whoall_' . $actions['action'] . '_' . $actions['sa']];
 			}
@@ -509,87 +456,4 @@ function determineActions($urls, $preferred_prefix = false)
 	{
 		return $data;
 	}
-}
-
-/**
- * Prepare credits for display.
- *
- * - This is a helper function, used by admin panel for credits and support page, and by the credits page.
- */
-function prepareCreditsData()
-{
-	global $txt;
-
-	$credits = array();
-
-	// Don't blink. Don't even blink. Blink and you're dead.
-	$credits['credits'] = array(
-		array(
-			'pretext' => $txt['credits_intro'],
-			'title' => $txt['credits_contributors'],
-			'groups' => array(
-				array(
-					'title' => $txt['credits_groups_contrib'],
-					'members' => array(
-						$txt['credits_contrib_list'],
-					),
-				),
-				array(
-					'title' => $txt['credits_groups_translators'],
-					'members' => array(
-						$txt['credits_translators_message'],
-					),
-				),
-			),
-		),
-	);
-
-	// Give credit to any graphic library's, software library's, plugins etc
-	$credits['credits_software_graphics'] = array(
-		'graphics' => array(
-			'<a href="https://icomoon.io">IcoMoon Free Icons</a> | These icons are licensed under <a href="https://creativecommons.org/licenses/by/4.0/">CC BY-SA 4.0</a>',
-			'<a href="https://github.com/googlefonts/noto-emoji">Noto Emoji</a> | &copy; Googlefonts | Licensed under <a href="https://www.apache.org/licenses/LICENSE-2.0">Apache License, Version 2.0</a>',
-			'<a href="https://openmoji.org">OpenMoji</a> | &copy; OpenMoji | Licensed under <a href="https://creativecommons.org/licenses/by-sa/4.0/">Creative Commons Attribution-ShareAlike 4.0</a>',
-			'<a href="http://www.oxygen-icons.org/">Oxygen Icons</a> | These icons are licensed under <a href="https://creativecommons.org/licenses/by-sa/3.0/">CC BY-SA 3.0</a>',
-			'<a href="https://github.com/twitter/twemoji">Twitter Emoji</a> | &copy; Twitter, Inc and other contributors | Licensed under <a href="https://github.com/twitter/twemoji/blob/master/LICENSE/">MIT</a>',		),
-		'fonts' => array(
-			'<a href="http://openfontlibrary.org/en/font/architect-s-daughter">Architect\'s Daughter</a> | &copy; 2010 <a href="http://kimberlygeswein.com/">Kimberly Geswein</a> | This font is licensed under the SIL Open Font License, Version 1.1',
-			'<a href="http://openfontlibrary.org/en/font/klaudia-and-berenika">Berenika</a> | &copy; 2011 wmk69 | This font is licensed under the SIL Open Font License, Version 1.1',
-			'<a href="http://openfontlibrary.org/en/font/dotrice">Dotrice</a> | &copy; 2010 <a href="http://hisdeedsaredust.com/">Paul Flo Williams</a> | This font is licensed under the SIL Open Font License, Version 1.1',
-			'<a href="http://fontawesome.io/">Font Awesome</a> | Created by Dave Gandy | This font is licensed under the SIL Open Font License, Version 1.1',
-			'<a href="http://openfontlibrary.org/en/font/press-start-2p">Press Start 2P</a> | &copy; 2012 Cody "CodeMan38" Boisclair | This font is licensed under the SIL Open Font License, Version 1.1',
-			'<a href="http://openfontlibrary.org/en/font/vds">VDS</a> | &copy; 2012 <a href="http://www.wix.com/artmake1/artmaker">artmaker</a> | This font is licensed under the SIL Open Font License, Version 1.1',
-			'<a href="http://openfontlibrary.org/en/font/vshexagonica-v1-0-1">vSHexagonica</a> | &copy; 2012 T.B. von Strong | This font is licensed under the SIL Open Font License, Version 1.1',
-		),
-		'software' => array(
-			'<a href="https://ichord.github.com/At.js">At.js</a> | &copy; Chord Luo | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="http://lab.ejci.net/favico.js/">Favico.js</a> | &copy; Miroslav Magda | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="https://code.google.com/p/google-code-prettify/">Google Code Prettify</a> | Licensed under <a href="https://opensource.org/licenses/Apache-2.0">Apache License, Version 2.0</a>',
-			'<a href="http://pajhome.org.uk/crypt/md5">Javascript Crypt</a> | &copy; Angel Marin, Paul Johnston | Licensed under <a href="https://opensource.org/licenses/BSD-3-Clause">The BSD License</a>',
-			'<a href="https://jquery.com/">JQuery</a> | &copy; jQuery Foundation and other contributors | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="https://jqueryui.com/">JQuery UI</a> | &copy; jQuery Foundation and other contributors | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="https://github.com/wikimedia/mediawiki-libs-Minify">JavaScriptMinifier</a> &copy Paul Copperman | Licensed under <a href="https://www.apache.org/licenses/LICENSE-2.0">Apache License, Version 2.0</a>',
-			'<a href="https://github.com/mailcheck">MailCheck</a> | &copy; Received Inc | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="https://github.com/mneofit/multiselect">Multiselect</a> | &copy; Mikhail Neofitov | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="https://github.com/michelf/php-markdown">PHP Markdown Lib</a> | &copy; Michel Fortin | Licensed under <a href="https://github.com/michelf/php-markdown/blob/lib/License.md">BSD-style open source</a>',
-			'<a href="https://github.com/ttsvetko/HTML5-Desktop-Notifications">Push.js</a> | &copy; Tyler Nickerson | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="https://www.sceditor.com/">SCEditor</a> | &copy; Sam Clarke | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="https://sourceforge.net/projects/simplehtmldom/">Simple HTML DOM</a> | Licensed under <a href="https://opensource.org/licenses/MIT">The MIT License (MIT)</a>',
-			'<a href="http://www.simplemachines.org/">Simple Machines</a> | &copy; Simple Machines | Licensed under <a href="https://www.simplemachines.org/about/smf/license.php">The BSD License</a>',
-			'<a href="https://github.com/tubalmartin/YUI-CSS-compressor-PHP-port">YUI-CSS compressor (PHP port)</a> | &copy; Yahoo! Inc | Licensed under <a href="https://opensource.org/licenses/BSD-3-Clause">The BSD License</a>',
-			'<a href="https://github.com/Frenzie/elk-quick-quote">Quick Quote</a> | &copy; Frans de Jonge | Licensed under <a href="https://opensource.org/licenses/BSD-3-Clause">The BSD License</a>',
-		),
-	);
-
-	// Add-ons authors: to add credits, the simpler and better way is to add in your package.xml the <credits> <license> tags.
-	// Support for addons that use the <credits> tag via the package manager
-	$credits['credits_addons'] = addonsCredits();
-
-	// An alternative for addons credits is to use a hook.
-	call_integration_hook('integrate_credits', array(&$credits));
-
-	// Copyright information
-	$credits['copyrights']['elkarte'] = '&copy; 2012 - ' . Util::strftime('%Y', time()) . ' ElkArte Forum contributors';
-
-	return $credits;
 }
